requirements.txt
certifi>=2024.0.0
charset-normalizer>=3.3.2
decorator>=5.1.1
idna>=3.6
imageio>=2.34.0
imageio-ffmpeg>=0.4.9
moviepy>=1.0.3
numpy>=2.2.0
Pillow>=10.3.0
proglog>=0.1.10
requests>=2.32.3
tqdm>=4.66.4
urllib3>=2.2.1
